test: add collector details > collect data for a specific account test case (#1578)
